low_level/screenshot-rs/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "screenshot_rs"
version = "0.1.0"
description = "Fast Wayland screenshot capture with numpy support"
requires-python = ">=3.8"
dependencies = ["numpy", "pillow"]

[tool.maturin]
features = ["pyo3/extension-module"]
